pyproject.toml
[project]
name = "bigomegaswork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright",
    "litellm",
    "python-dotenv",
]
